refactor: organize gemini prompts into seperate files
